New changes for pre-release.
Added further visual styling to the inspector. Added section headers.
Improved the performance of setting definition loading.
Added support for combining the 1st and 2nd shading tones.
Detail emission properties are now in a seperate category for future expansion.
The appearance of cutout dithering will now take MSAA level into account.
Properties that are initialised with a black tint colour that multiplies the texture will now reset the tint colour when a texture is assigned. (i.e. assigning a tint texture will reset the tint colour if it's black)