New changes for pre-release.
Added material baking. When baked, materials can't be modified, but they'll have a much smaller build size, and they'll also be a bit cheaper to render too. Baked materials will place a copy of the shader files in the same folder as the material, with the material settings baked in.
Fixed an issue where single-pass stereo instancing rendered incorrectly.
Fixed an issue where the shader failed to compile correctly in Unity 2019.4.
Matcap tint properties will now be transferred from UCTS.