New changes for pre-release.
Added a fix for a compile error that appeared in certain conditions where screen-space shadows were active but the keyword set was for shadowmask.
Baked shaders will flatten paths to avoid an issue where the result created a path that was too long.
Adjusted appearance of the material baking button.
Added a tentative fix for #1.