New changes for pre-release.
Fixed an issue where, when a material was baked and un-baked, the correct specular keyword was not assigned.
Fixed an issue where animated materials would not be reflected in the cast shadow.
Fixed an issue where outlines would incorrectly receive the effect of ambient rim-lighting.
Fixed an issue where animation did not display correctly.
Added more graceful handling to the inspector for cases where default properties do not exist.
Reorganised some internal functions for future extension.