New changes for pre-release.
Fixed the "Bake materials" button having text that was too long when selecting multiple materials
The "Disable Dithering" setting has been replaced with a dropdown for "Transparency Mode" for clarity
Added the ability to change the scale and offset settings for the clipping mask when it's active.
Fixed an issue where detail maps behaved incorrectly when using tone seperation.