New changes for pre-release.
Add Outline Depth Offset option. This allows you to push and pull outlines relative to the camera position, allowing you to set hide outlines when one object overlaps another, without needing to write it into the vertex colours.
Reworked outline generation in preparation for a future feature.
Improved the functioning of the inventory settings to fix an exception that occurred on shaders baked with an older version.
Rewrote the material transfer system. Some options are now handled that weren't before. It'll also be easier to add future support for other properties.
Fixed an issue where alpha to coverage transparency was too transparent compared to regular transparency.
Fixed an issue with the realtime light clamp where vertex lights were not clamped.
Fixed an issue with the realtime light clamp where the clamped value was brighter than the unclamped value in certain conditions.
Fixed the issue where changing shader from an outline variant to a non-outline variant, or vice-versa, would immediately be reverted by the inspector GUI.